+++ td/telegram/ContactsManager.h
@@ -532,23 +532,23 @@ class ContactsManager final : public Actor {
                                 string additional_query, int32 offset, int32 limit, int32 additional_limit,
                                 Promise<DialogParticipants> &&promise);
 
-  int32 get_user_id_object(UserId user_id, const char *source) const;
+  int64 get_user_id_object(UserId user_id, const char *source) const;
 
   tl_object_ptr<td_api::user> get_user_object(UserId user_id) const;
 
-  vector<int32> get_user_ids_object(const vector<UserId> &user_ids, const char *source) const;
+  vector<int64> get_user_ids_object(const vector<UserId> &user_ids, const char *source) const;
 
   tl_object_ptr<td_api::users> get_users_object(int32 total_count, const vector<UserId> &user_ids) const;
 
   tl_object_ptr<td_api::userFullInfo> get_user_full_info_object(UserId user_id) const;
 
-  int32 get_basic_group_id_object(ChatId chat_id, const char *source) const;
+  int64 get_basic_group_id_object(ChatId chat_id, const char *source) const;
 
   tl_object_ptr<td_api::basicGroup> get_basic_group_object(ChatId chat_id);
 
   tl_object_ptr<td_api::basicGroupFullInfo> get_basic_group_full_info_object(ChatId chat_id) const;
 
-  int32 get_supergroup_id_object(ChannelId channel_id, const char *source) const;
+  int64 get_supergroup_id_object(ChannelId channel_id, const char *source) const;
 
   tl_object_ptr<td_api::supergroup> get_supergroup_object(ChannelId channel_id) const;
 
@@ -1335,7 +1335,7 @@ class ContactsManager final : public Actor {
 
   int32 get_user_was_online(const User *u, UserId user_id) const;
 
-  int32 get_contacts_hash();
+  int64 get_contacts_hash();
 
   void update_contacts_hints(const User *u, UserId user_id, bool from_database);
 
@@ -1435,7 +1435,7 @@ class ContactsManager final : public Actor {
   void on_load_administrator_users_finished(DialogId dialog_id, vector<DialogAdministrator> administrators,
                                             Result<> result, Promise<Unit> promise);
 
-  void reload_dialog_administrators(DialogId dialog_id, int32 hash, Promise<Unit> &&promise);
+  void reload_dialog_administrators(DialogId dialog_id, int64 hash, Promise<Unit> &&promise);
 
   void remove_dialog_suggested_action(SuggestedAction action);
