+++ td/telegram/Global.h
@@ -338,10 +338,10 @@ class Global final : public ActorContext {
     return parameters_;
   }
 
-  int32 get_my_id() const {
+  int64 get_my_id() const {
     return my_id_;
   }
-  void set_my_id(int32 my_id) {
+  void set_my_id(int64 my_id) {
     my_id_ = my_id;
   }
 
@@ -470,7 +470,7 @@ class Global final : public ActorContext {
 
   unique_ptr<ConfigShared> shared_config_;
 
-  int32 my_id_ = 0;  // hack
+  int64 my_id_ = 0;  // hack
 
   static int64 get_location_key(double latitude, double longitude);
