+++ td/telegram/LinkManager.h
@@ -58,10 +58,10 @@ class LinkManager final : public Actor {
 
   void get_external_link_info(string &&link, Promise<td_api::object_ptr<td_api::LoginUrlInfo>> &&promise);
 
-  void get_login_url_info(FullMessageId full_message_id, int32 button_id,
+  void get_login_url_info(FullMessageId full_message_id, int64 button_id,
                           Promise<td_api::object_ptr<td_api::LoginUrlInfo>> &&promise);
 
-  void get_login_url(FullMessageId full_message_id, int32 button_id, bool allow_write_access,
+  void get_login_url(FullMessageId full_message_id, int64 button_id, bool allow_write_access,
                      Promise<td_api::object_ptr<td_api::httpUrl>> &&promise);
 
   void get_link_login_url(const string &url, bool allow_write_access,
