+++ td/telegram/ReplyMarkup.h
@@ -42,7 +42,7 @@ struct InlineKeyboardButton {
     CallbackWithPassword
   };
   Type type;
-  int32 id = 0;  // UrlAuth only, button_id or (2 * request_write_access - 1) * bot_user_id
+  int64 id = 0;  // UrlAuth only, button_id or (2 * request_write_access - 1) * bot_user_id
   string text;
   string forward_text;  // UrlAuth only
   string data;
