+++ td/telegram/net/NetQuery.h
@@ -325,7 +325,7 @@ class NetQuery final : public TsListNode<NetQueryDebug> {
     source_ = std::move(source);
   }
 
-  static int32 get_my_id();
+  static int64 get_my_id();
 
   static int32 tl_magic(const BufferSlice &buffer_slice);
