+++ test/secret.cpp
@@ -85,7 +85,7 @@ class InputUser {
 
 class inputUser final : public InputUser {
  public:
-  int32 user_id_{};
+  int64 user_id_{};
   int64 access_hash_{};
 
   static const int32 ID = -668391402;
@@ -254,14 +254,14 @@ class encryptedChat final {
   int32 id_{};
   int64 access_hash_{};
   int32 date_{};
-  int32 admin_id_{};
-  int32 participant_id_{};
+  int64 admin_id_{};
+  int64 participant_id_{};
   BufferSlice g_a_or_b_;
   int64 key_fingerprint_{};
 
   encryptedChat() = default;
 
-  encryptedChat(int32 id_, int64 access_hash_, int32 date_, int32 admin_id_, int32 participant_id_,
+  encryptedChat(int32 id_, int64 access_hash_, int32 date_, int64 admin_id_, int64 participant_id_,
                 BufferSlice &&g_a_or_b_, int64 key_fingerprint_)
       : id_(id_)
       , access_hash_(access_hash_)
@@ -763,8 +763,8 @@ class Master final : public Actor {
     auto old_context = set_context(std::make_shared<Global>());
     alice_ = create_actor<SecretChatProxy>("SecretChatProxy alice", "alice", actor_shared(this, 1));
     bob_ = create_actor<SecretChatProxy>("SecretChatProxy bob", "bob", actor_shared(this, 2));
-    send_closure(alice_->get_actor_unsafe()->actor_, &SecretChatActor::create_chat, UserId(2), 0, 123,
-                 PromiseCreator::lambda([actor_id = actor_id(this)](Result<SecretChatId> res) {
+    send_closure(alice_->get_actor_unsafe()->actor_, &SecretChatActor::create_chat, UserId(static_cast<int64>(2)), 0,
+                 123, PromiseCreator::lambda([actor_id = actor_id(this)](Result<SecretChatId> res) {
                    send_closure(actor_id, &Master::got_secret_chat_id, std::move(res), 0);
                  }));
   }
@@ -869,7 +869,7 @@ class Master final : public Actor {
     send_message(2, "appo");
     set_timeout_in(1);
   }
-  void send_ping(int id, int cnt) {
+  void send_ping(int32 id, int cnt) {
     if (cnt % 200 == 0) {
       LOG(ERROR) << "Send ping " << tag("id", id) << tag("cnt", cnt);
     } else {
@@ -878,7 +878,7 @@ class Master final : public Actor {
     string text = PSTRING() << "PING: " << cnt;
     send_message(id, std::move(text));
   }
-  void send_message(int id, string text) {
+  void send_message(int32 id, string text) {
     auto random_id = Random::secure_int64();
     LOG(INFO) << "Send message: " << tag("id", id) << tag("text", text) << tag("random_id", random_id);
     sent_messages_[random_id] = Message{id, text};
